requirements.txt
allure-pytest==2.13.2
pytest==7.4.3
selenium==4.16.0